INTEGRATION: the world-turn pipeline runs end-to-end (+2 real rules bugs fixed)
New runtime/game_turn.c wires the per-year world step into the shell's
end_turn: per power - census rebuild (func_042138) -> colony production +
SoL/Tory (func_02D658) -> Europe market drift (func_0305A8, silent) ->
crown sentiment/REF buildup (func_03E162); then turn/year advance.
--smoke=N runs it headless on a synthetic fixture (no game data needed):

  SMOKE PASS: 60 world turns, year 1552
  REF grew to 4 units (reg 3 cav 1) - matching the byte-verified
  42/turn sentiment at Viceroy difficulty and the 1800/unit arm ratios.

Integration testing caught and fixed TWO real rules-layer bugs:
1. sol_tory.c: phases 1+2 are ONE INTERLEAVED LOOP in the original
   (entry @0x2D6D5 jumps to the shared compare; the pressure body's
   branches reach the settle body at 0x2D6D8 / the shared tail at 0x2D877;
   de-Witt path's 0x181F:0xCFE result re-enters as `over` via the
   JMP 0x2D6E0 stack-cleanup trick).  The previous two-sequential-loops
   port made phase-2's goto re-enter a foreign loop -> infinite, AND ran
   the settle body unconditionally.  Re-ported byte-anchored.
2. production_support.c: the building prereq chain byte is SIGNED
   (CWDE @0x8679); the uint8_t read made 0xFF terminators read as 255
   -> infinite 0<->255 chain walks.

Modernizations: market_set_active writes DGROUP [0x9E12]/[0x84FC] + host
g_market ptr; euro-supply/gold dword reads are int32_t (not 64-bit long);
real bodies for revolution_flag_5382 / king_sentiment_accum_22; linkfloor
extras alias the named market globals (price_seed, display_price, ...).

+++ viceroy_source/src/colony/production_support.c
@@ -75,7 +75,7 @@ extern uint16_t g_8DC6;                   /* DGROUP:0x8DC6 */
  * @asm 0x8674/0x86AF/0x86DA/0x870D: MOV al, byte ptr [bx - 0x707a] with
  *      bx = (idx*3)<<2 = idx*12, and (0x10000-0x707a)=0x8F86.
  * Declared in globals.h as g_table_8F86_stride12[]. */
-#define g_chain_next(idx)  (g_table_8F86_stride12[(idx) * 12])
+#define g_chain_next(idx)  ((int8_t)g_table_8F86_stride12[(idx) * 12])  /* SIGNED per @asm 0x8679 CWDE */
 
 /* signed-byte table at DGROUP:0x2F4 — per-good "chain-start building id".
  * @asm 0x8DAE: MOV al, byte ptr [bx + 0x2f4]   (bytes 8A 87 F4 02) */
